% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-package.R
\docType{package}
\name{interpElections-package}
\alias{interpElections}
\alias{interpElections-package}
\title{interpElections: Spatial Interpolation of Electoral Data via Inverse Distance Weighting}
\description{
Spatial interpolation of electoral data via inverse distance weighting
with per-zone optimized decay parameters. Designed for disaggregating
voting results from polling locations into census tracts using
travel-time-based IDW, with optional GPU acceleration via torch.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://antrologos.github.io/interpElections/}
  \item \url{https://github.com/antrologos/interpElections}
  \item Report bugs at \url{https://github.com/antrologos/interpElections/issues}
}

}
\author{
\strong{Maintainer}: Rogerio Barbosa \email{rogerio.barbosa@iesp.uerj.br}

Authors:
\itemize{
  \item Lucas Gelape
}

}
\keyword{internal}
