% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result-methods.R
\name{plot.interpElections_result}
\alias{plot.interpElections_result}
\title{Plot an interpolated variable on a map}
\usage{
\method{plot}{interpElections_result}(x, var = NULL, ...)
}
\arguments{
\item{x}{An \code{interpElections_result} object.}

\item{var}{Character. Name of the variable to plot. Must be one of
\code{x$interp_cols}. If NULL, plots the first interpolated variable.}

\item{...}{Additional arguments passed to \link[sf:plot]{plot.sf()}.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Produces a quick choropleth map of an interpolated variable using
the \code{tracts_sf} stored in the result object.
}
