% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-checks.R
\name{print.interpElections_diagnostics}
\alias{print.interpElections_diagnostics}
\title{Print method for interpElections_diagnostics}
\usage{
\method{print}{interpElections_diagnostics}(x, ...)
}
\arguments{
\item{x}{An \code{interpElections_diagnostics} object.}

\item{...}{Ignored.}
}
\description{
Print method for interpElections_diagnostics
}
